requirements.txt
association-measures==0.1.3
certifi==2018.4.16
cffi==1.14.0
chardet==3.0.4
ckip-segmenter==1.0.2
cryptography==2.8
Cython==0.29.15
cwb-ccc==0.9.6
defusedxml==0.5.0
Django==2.0.7
django-allauth==0.41.0
django-material==1.3.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
idna==2.7
jseg==0.0.4
nltk==3.0.3
numpy==1.18.2
oauthlib==2.1.0
pandas==1.0.3
pycparser==2.20
PyJWT==1.6.4
python-dateutil==2.8.1
python-memcached==1.59
python-social-auth==0.3.6
python3-openid==3.1.0
pytz==2018.5
requests==2.19.1
requests-oauthlib==1.0.0
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==3.3.0
Unidecode==1.1.1
urllib3==1.23
uWSGI==2.0.17.1
